move margin test to simple_test file
